replaced stb image lib with FreeImage, added (experimental yet) app for dealing with effect .str files, and a few minor changes

+++ src/format/Act.hpp
@@ -55,14 +55,14 @@ struct Act {
     /**
      * Loads from memory buffer.
      *
-     * @throws InvalidFile on failure.
+     * @throws InvalidResource on failure.
      */
     void load(const Buffer& buf);
 
     /// Saves to memory buffer.
     void save(Buffer& buf) const;
 
-    struct { int major, minor; } version;
+    struct { uint8_t major, minor; } version;
     std::vector<Animation> animations;
     std::vector<Sound> sounds;
 };
